Move skip button from grading buttons to context menu on study page
This gives more space to the Correct/Incorrect buttons by moving
the less frequently used Skip action into the card actions menu.
