+++ client/src/app/shared/card-grading-buttons/card-grading-buttons.component.html
@@ -1,8 +1,4 @@
 <div class="grading-buttons">
-  <button mat-raised-button class="grade-skip" (click)="skipCard()">
-    <mat-icon>skip_next</mat-icon>
-    Skip
-  </button>
   <button mat-raised-button class="grade-incorrect" (click)="gradeCard('Again')">
     <mat-icon>close</mat-icon>
     Incorrect
