+++ client/src/app/study/learn-card/learn-card.component.html
@@ -97,7 +97,6 @@
   <div class="grading-buttons-container" [class.hidden]="!isRevealed()">
     <app-card-grading-buttons
       [card]="cardResource"
-      [sourceId]="sourceId()"
       [learningPartnerId]="currentTurn().partnerId"
       [reviewDuration]="reviewDuration()"
       (cardProcessed)="onCardProcessed()">
@@ -105,6 +104,7 @@
     <div class="grading-row-actions">
       <app-card-actions
         [card]="cardResource"
+        [sourceId]="sourceId()"
         [languageTexts]="languageTexts()"
         (cardProcessed)="onCardProcessed()">
       </app-card-actions>
